backend/requirements.txt
google-generativeai
# Not sure about the above one
pandas
opencv-python
numpy
scikit-learn
tensorflow
requests
Pillow
transformers
opencv-python
numpy
torch
python-dotenv
# yolov4.weights
# yolov4.cfg
# coco.names
